+++ src/index.jsx
@@ -18,8 +18,10 @@ import { useState, useEffect, useRef, createContext, useContext } from 'react';
  * @property {QueryResultCache} cache
  * 
  * @typedef {object} DefaultUseQueryOptions
+ * @property {string} [name] 
  * @property {string} [selector] 
- * @property {function} [manipulate] 
+ * @property {array|function} [pick]
+ * @property {function} [transform] 
  * @property {string} [method="GET"] 
  * @property {object} [headers] 
  * @property {*} [body] 
@@ -28,8 +30,10 @@ import { useState, useEffect, useRef, createContext, useContext } from 'react';
  */
 
 const DEFAULT_OPTIONS = {
+    name: undefined,
     selector: undefined,
-    manipulate: undefined,
+    transform: undefined,
+    pick: undefined,
     method: "GET",
     headers: undefined,
     body: undefined,
@@ -65,6 +69,9 @@ const validateOptions = (options) => {
     if (options.selector !== undefined && typeof options.selector !== "string") {
         logDebugger("Validation error for `options.selector`", isDebuggerActivated, true, new Error("Selector must be a string. It should contains key value to select from result object."));
         return false;
+    } else if (options.name !== undefined && typeof options.name !== "string") {
+        logDebugger("Validation error for `options.name`", isDebuggerActivated, true, new Error("Name must be a string. It should contains data name to retrieve it from state."));
+        return false;
     } else if (options.method !== undefined && (typeof options.method !== "string" || ["GET", "POST", "PUT", "PATCH", "DELETE"].indexOf(options.method.toUpperCase()) === -1)) {
         logDebugger("Validation error for `options.method`", isDebuggerActivated, true, new Error('Method must be a string. It should be one of "GET", "POST", "PUT", "PATCH", "DELETE".'));
         return false;
@@ -74,8 +81,11 @@ const validateOptions = (options) => {
     } else if (options.isDebuggerActivated !== undefined && typeof options.isDebuggerActivated !== "boolean" && Array.isArray(options.headers)) {
         logDebugger("Validation error for `options.isDebuggerActivated`", true, true, new Error("isDebuggerActivated must be a boolean. It should be activated if you need to debug all process."));
         return false;
-    } else if (options.manipulate !== undefined && typeof options.manipulate !== "function") {
-        logDebugger("Validation error for `options.manipulate`", true, true, new Error("manipulate must be a function. It rappresent the funcion to manipulate data before saving on state."));
+    } else if (options.transform !== undefined && typeof options.transform !== "function") {
+        logDebugger("Validation error for `options.transform`", true, true, new Error("transform must be a function. It rappresent the function to transform data before saving on state."));
+        return false;
+    } else if (options.pick !== undefined && (typeof options.pick !== "function" && !Array.isArray(options.pick))) {
+        logDebugger("Validation error for `options.pick`", true, true, new Error("pick must be a function or an array. It rappresent the function or the array to pick just a portion of data."));
         return false;
     } else if (options.cacheTimeout !== undefined && typeof options.cacheTimeout !== "number") {
         logDebugger("Validation error for `options.cacheTimeout`", true, true, new Error("cacheTimeout must be a number. It rappresent the timeout to remove cached data from memory in milliseconds."));
@@ -116,14 +126,24 @@ export const QueryProvider = ({ children }) => {
     )
 }
 
-export const useQueryContext = () => useContext(QueryContext);
+export const useQueryContext = (name = undefined) => {
+    const [data, setData] = useContext(QueryContext);
+    
+    if (name) { 
+        return [data ? data[name] : data, (value) => setData((prev) => ({ ...prev, [name]: value }))];
+    } else {
+        return [data, setData];
+    }
+};
 
 /**
  * Use query hook that manage requests, cache and other features.
  * @param {string} url 
  * @param {object} options
+ * @param {string} [options.name] Name must be a string. It should contains data state name.
  * @param {string} [options.selector] Selector must be a string. It should contains key value to select from result object.
- * @param {function} [options.manipulate] manipulate must be a function. It rappresent the funcion to manipulate data before saving on state. If you need to mutate data, use `mutate` function instead
+ * @param {string} [options.pick] Pick must be a function or an array of string. It rappresent the function or the array to pick just a portion of data.
+ * @param {function} [options.transform] Transform must be a function. It rappresent the funcion to transform data before saving on state. If you need to mutate data, use `mutate` function instead
  * @param {string} [options.method] Method must be a string. It should be one of "GET", "POST", "PUT", "PATCH", "DELETE".
  * @param {object} [options.headers] Headers must be an object. It should contains request headers key value.
  * @param {*} [options.body]
@@ -132,13 +152,13 @@ export const useQueryContext = () => useContext(QueryContext);
  * @returns {QueryResult}
  */
 export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
-    const { selector, manipulate, method, headers, body, isDebuggerActivated } = { ...DEFAULT_OPTIONS, ...options };
+    const { name, selector, pick, transform, method, headers, body, isDebuggerActivated } = { ...DEFAULT_OPTIONS, ...options };
 
     validateOptions({ ...DEFAULT_OPTIONS, ...options });
     
     const cache = useRef({});
 
-    const [data, setData] = useQueryContext();
+    const [data, setData] = useQueryContext(name);
     const [error, setError] = useState(null);
     const [loading, setLoading] = useState(false);
 
@@ -150,7 +170,8 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
 
         if (cache.current[url]) {
             logDebugger("Get data from cache, no need a new request", isDebuggerActivated);
-            if (typeof manipulate === "function") logDebugger("Manipulate data before saving", isDebuggerActivated);
+            if (typeof transform === "function") logDebugger("transform data before saving", isDebuggerActivated);
+            if (typeof pick === "function" || Array.isArray(pick)) logDebugger("pick data before saving", isDebuggerActivated);
 
             setData(cache.current[url]);
             timeoutCacheClear(options, cache);
@@ -166,10 +187,15 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
 
             if (response.ok) {
                 const _result = selector ? result[selector] : result;
-                if (typeof manipulate === "function") logDebugger("Manipulate data before saving", isDebuggerActivated);
-                setData(typeof manipulate === "function" ? manipulate(_result) : _result);
+                if (typeof transform === "function") logDebugger("transform data before saving", isDebuggerActivated);
+                let _transformedData = typeof transform === "function" ? transform(_result) : _result;
+                if (typeof pick === "function" || Array.isArray(pick)) {
+                    logDebugger("pick data before saving", isDebuggerActivated);
+                    _transformedData = JSON.parse(JSON.stringify(_transformedData, pick));
+                }
+                setData(_transformedData);
                 logDebugger("Request done", isDebuggerActivated);
-                cache.current[url] = typeof manipulate === "function" ? manipulate(_result) : _result;
+                cache.current[url] = _transformedData;
                 timeoutCacheClear(options, cache);
                 logDebugger(`New data saved on cache for: "${url}"`, isDebuggerActivated);
             } else {
@@ -193,7 +219,7 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
         data,
         error,
         loading,
-        updateData: setData,
+        mutate: setData,
         refresh: fetchData,
         cache: {
             get: (url) => {