+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] 2024-03-13
+
+Added `name` option to get granular control over state context
+
+## [1.0.9] 2024-03-13
+
+Renamed option `manipulate` to `transform` && added `pick` option.
+
 ## [1.0.8] 2024-03-01
 
 Fixed bug on bundle process that broke`import` inside react projects.